Import Geant4 9.5.0 source tree

+++ examples/advanced/medical_linac/src/DicomPhantomParameterisationColour.cc
@@ -98,7 +98,7 @@ ComputeMaterial(const G4int copyNo, G4VPhysicalVolume * physVol, const G4VToucha
   if( physVol ) {
     G4String mateName = mate->GetName();
     unsigned int iuu = mateName.find("__");
-    if( iuu != std::string::npos ) {
+    if( iuu != (unsigned int) std::string::npos ) {
       mateName = mateName.substr( 0, iuu );
     }
     std::map<G4String,G4VisAttributes*>::const_iterator ite = fColours.find(mateName);
